game-backend/pyproject.toml
[tool.poetry]
name = "game-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "game_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.0"
python-dotenv = "^1.0.0"
black = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
